requirements.txt
torch>=2.0
torchvision
timm>=0.9.0
transformers>=4.30.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
pillow>=10.0.0
pyyaml>=6.0
tqdm>=4.65.0
grad-cam>=1.4.8
opencv-python>=4.8.0
